重构addon types
