+++ web/src/util/useAddon/addon.ts
@@ -2,12 +2,13 @@ import Bowser from 'bowser';
 
 import type { Message, MessageRequest, MessageResponse } from './msg';
 import { MessageRequestType, MessageResponseType } from './msg';
-import type {
-  ClientMethods,
-  InfoResult,
-  JobNewResult,
-  SerializableRequest,
-  SerializableResponse,
+import {
+  Response2SerResp,
+  serializeRequest,
+  type ClientMethods,
+  type InfoResult,
+  type JobNewResult,
+  type SerializableResponse,
 } from './types';
 
 async function sendMessageChrome<T>(msg: Message): Promise<T> {
@@ -72,71 +73,6 @@ function createAddonApi() {
 let msgId = 1;
 
 const api = createAddonApi();
-
-async function Response2SerResp(
-  response: Response,
-): Promise<SerializableResponse> {
-  const headers: [string, string][] = Array.from(response.headers.entries());
-  const bodyText = await response.text();
-
-  const serializableResponse = {
-    body: bodyText,
-    status: response.status,
-    statusText: response.statusText,
-    ok: response.ok,
-    headers: headers,
-    redirected: response.redirected,
-    url: response.url,
-    type: response.type,
-  };
-  return serializableResponse;
-}
-
-export async function serializeRequest(
-  request: RequestInfo,
-): Promise<SerializableRequest | string> {
-  if (typeof request === 'string') {
-    return request;
-  }
-
-  const headers: [string, string][] = Array.from(request.headers.entries());
-  console.log('serializeRequest: ', headers);
-
-  const req: SerializableRequest = {
-    url: request.url,
-    method: request.method,
-    headers,
-    body: request.body ? await request.text() : undefined,
-    mode: request.mode,
-    credentials: request.credentials,
-    cache: request.cache,
-    redirect: request.redirect,
-    referrer: request.referrer,
-    integrity: request.integrity,
-  };
-  return req;
-}
-
-export function deserializeRequest(req: SerializableRequest): RequestInfo {
-  if (typeof req === 'string') {
-    return req;
-  }
-
-  console.log('deserializeRequest: ', req);
-  const init: RequestInit = {
-    method: req.method,
-    headers: new Headers(req.headers),
-    body: req.body,
-    mode: req.mode,
-    credentials: req.credentials,
-    cache: req.cache,
-    redirect: req.redirect,
-    referrer: req.referrer,
-    integrity: req.integrity,
-  };
-
-  return new Request(req.url, init);
-}
 
 export class AddonClient {
   jobId: string | undefined;
