+++ web/src/util/useAddon/types.ts
@@ -23,6 +23,71 @@ export interface SerializableRequest {
   integrity?: string;
 }
 
+export async function Response2SerResp(
+  response: Response,
+): Promise<SerializableResponse> {
+  const headers: [string, string][] = Array.from(response.headers.entries());
+  const bodyText = await response.text();
+
+  const serializableResponse = {
+    body: bodyText,
+    status: response.status,
+    statusText: response.statusText,
+    ok: response.ok,
+    headers: headers,
+    redirected: response.redirected,
+    url: response.url,
+    type: response.type,
+  };
+  return serializableResponse;
+}
+
+export async function serializeRequest(
+  request: RequestInfo,
+): Promise<SerializableRequest | string> {
+  if (typeof request === 'string') {
+    return request;
+  }
+
+  const headers: [string, string][] = Array.from(request.headers.entries());
+  console.log('serializeRequest: ', headers);
+
+  const req: SerializableRequest = {
+    url: request.url,
+    method: request.method,
+    headers,
+    body: request.body ? await request.text() : undefined,
+    mode: request.mode,
+    credentials: request.credentials,
+    cache: request.cache,
+    redirect: request.redirect,
+    referrer: request.referrer,
+    integrity: request.integrity,
+  };
+  return req;
+}
+
+export function deserializeRequest(req: SerializableRequest): RequestInfo {
+  if (typeof req === 'string') {
+    return req;
+  }
+
+  console.log('deserializeRequest: ', req);
+  const init: RequestInit = {
+    method: req.method,
+    headers: new Headers(req.headers),
+    body: req.body,
+    mode: req.mode,
+    credentials: req.credentials,
+    cache: req.cache,
+    redirect: req.redirect,
+    referrer: req.referrer,
+    integrity: req.integrity,
+  };
+
+  return new Request(req.url, init);
+}
+
 export type InfoResult = {
   version: string;
 };
